Create Sam Rostami
Classification of water masses plays a fundamental role in the understanding of the relationship between ocean circulation and climate change. Also, it helps to understand the global ocean circulation, 
as well as influencing poleward heat transport, the export of salt, and nutrient distributions. This study is focusing on Cretaceous time about 66 million years ago. 
There were different connections to other basins compared to today, with enhanced connections to east-west flow in the equatorial region.  In addition, PCO2 estimates of 530-620 ppm.
This study combines data for these chemical and biological proxies to reconstruct water mass sources, flow paths, and evolution of circulation patterns in the North Atlantic 
during the Campanian to Maastrichtian (80-60 million years ago) to advance our understanding of the relationship between deep ocean circulation and climate at this time.
Therefore, I used the Azure machine learning platform and compare four different algorithms to train my data and achieve the best model for the prediction of different water masses in the North Atlantic.